Update credentials.yml on configuration & password change
